# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion",
        "description": "Documentation de l'API",
        "contact": {
            "email": "support@tonapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Liste des catégories"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une nouvelle catégorie",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Informatique"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée avec succès"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Afficher une catégorie spécifique",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la catégorie"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie",
                "operationId": "0e686b2748211cc688333ed705dc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nouvelles technologies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "4c12e22a7f8c617bd83bfa1fdc05428d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Catégories",
            "description": "Gestion des catégories"
        }
    ]
}